Add option (-s) to explicitly specify the URL scheme
Explicitly specifying the URL scheme might be required depending on how the PAC file is written. For example, if the script returns a different value based on the URL scheme then passing the full URL to the `FindProxyForURL` method might be required.

Here's an example of such a (real life) PAC script:

```js
if ((url.substring(0,5) != "http:") && (url.substring(0,6) != "https:"))
return "DIRECT";
```

Since the [request URL scheme is not known][1], passing it explicitly through a command line option solves this issue.

The README has also been updated to reflect the intended usage of the new `-s` option.

Please bear in mind that this is the first time I ever write some go code. 😇

[1]: https://github.com/golang/go/issues/28940